+++ framework/inputbuf.h
@@ -38,11 +38,11 @@ struct InputBuf {
     int handling_msg;       /* Channel in the process of handling a message */
 #if ENABLE_ZeroCopy
     int out_size_mode;      /* Checking the binary data size */
-    int out_data_size;      /* Size of the bin data to get */
+    size_t out_data_size;   /* Size of the bin data to get */
     int inp_size_mode;      /* (Read done) Checking the binary data size */
-    int inp_data_size;      /* (Read done) Size of the bin data to get */
+    size_t inp_data_size;   /* (Read done) Size of the bin data to get */
 #endif
-    void (*post_read)(InputBuf *, unsigned char *, int);
+    void (*post_read)(InputBuf *, unsigned char *, size_t);
     void (*wait_read)(InputBuf *);
     void (*trigger_message)(InputBuf *);
 };
@@ -57,7 +57,7 @@ extern void ibuf_init(InputBuf * ibuf, InputStream * inp);
 extern void ibuf_trigger_read(InputBuf * ibuf);
 extern int ibuf_get_more(InputBuf * ibuf, int peeking);
 extern void ibuf_flush(InputBuf * ibuf);
-extern void ibuf_read_done(InputBuf * ibuf, int len);
+extern void ibuf_read_done(InputBuf * ibuf, size_t len);
 extern int ibuf_start_message(InputBuf * ibuf);
 
 #endif /* D_input_buf */
